# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "this is a service that helps customers order items",
        "title": "Client Orders",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "maxwellgithinji@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/maxwellgithinji/customer_orders/blob/develop/LICENSE"
        },
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/delete/item/{id}": {
            "delete": {
                "description": "deletes an item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "enables a user to delete an item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Item Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/auth/item": {
            "post": {
                "description": "creates a new item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "creates a new item",
                "parameters": [
                    {
                        "description": "Item",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ItemPost"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ItemPost"
                        }
                    }
                }
            }
        },
...
        "/auth/orders": {
            "get": {
                "description": "Get all orders in the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Get all orders in the database",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Order"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "creates a new order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "creates a new order",
                "parameters": [
                    {
                        "description": "Order",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.OrderPost"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.OrderPost"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ItemPost": {
            "type": "object",
            "properties": {
                "item": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                }
            }
        },
...
        "models.Order": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "customer_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "item_id": {
                    "type": "integer"
                },
                "order_date": {
                    "type": "string"
                },
                "total_price": {
                    "type": "number"
                }
            }
        },
        "models.OrderPost": {
            "type": "object",
            "properties": {
                "item_id": {
                    "type": "integer"
                }
            }
        }
    }
}